moved server compilation out of compilerServer
