+++ backend/internal/compilerServer/daemon.go
@@ -103,11 +103,8 @@ func (app *CompiledApp) setupJsDaemon(processConfig *process.ProcessConfig) erro
 		return fmt.Errorf("failed to start app server: %w", err)
 	}
 
-	// Build the server bundle, if not already built
-	if app.ServerJs == "" {
-		if err := app.buildServerBundle(); err != nil {
-			return fmt.Errorf("failed to build app server: %w", err)
-		}
+	if err := app.buildServerBundle(); err != nil {
+		return fmt.Errorf("failed to build app server: %w", err)
 	}
 
 	// Figure out asset paths
